♻️ refactor: notification 위치 이동
참고용 주석 추가
