+++ .idea/workspace.xml
@@ -6,9 +6,9 @@
   <component name="ChangeListManager">
     <list default="true" id="6259cfc3-7d2e-4e15-bddb-98b070861ccc" name="Changes" comment="🩹 fix: createResponse.js 동작 수정">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/classes/models/game.class.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/classes/models/game.class.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/handler/index.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/handler/index.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/utils/notification/matchStart.notification.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/utils/notification/matchStart.notification.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/handler/user/matchRequest.handler.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/handler/user/matchRequest.handler.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/utils/notification/game.notification.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/handler/notification/game.notification.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/utils/notification/matchStart.notification.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/handler/notification/matchStart.notification.js" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
